+++ skforecast/model_selection/_utils.py
@@ -18,7 +18,7 @@ from sklearn.exceptions import NotFittedError
 
 from ..exceptions import IgnoredArgumentWarning
 from ..metrics import add_y_train_argument, _get_metric
-from ..utils import check_interval
+from ..utils import check_interval, date_to_index_position
 
 
 def initialize_lags_grid(
@@ -303,23 +303,39 @@ def check_backtesting_input(
     if forecaster_name == "ForecasterEquivalentDate" and isinstance(
         forecaster.offset, pd.tseries.offsets.DateOffset
     ):
+        # NOTE: Checks when initial_train_size is not None cannot be done here
+        # because the forecaster is not fitted yet and we don't know the
+        # window_size since pd.DateOffset is not a fixed window size.
         if initial_train_size is None:
             raise ValueError(
                 f"`initial_train_size` must be an integer greater than "
                 f"the `window_size` of the forecaster ({forecaster.window_size}) "
-                f"and smaller than the length of `{data_name}` ({data_length})."
+                f"and smaller than the length of `{data_name}` ({data_length}) or "
+                f"a date within this range of the index."
             )
     elif initial_train_size is not None:
+        if forecaster_name in forecasters_uni:
+            index = cv._extract_index(y)
+        else:
+            index = cv._extract_index(series)
+
+        initial_train_size = date_to_index_position(
+                                 index        = index, 
+                                 date_input   = initial_train_size, 
+                                 method       = 'validation',
+                                 date_literal = 'initial_train_size'
+                             )
         if initial_train_size < forecaster.window_size or initial_train_size >= data_length:
             raise ValueError(
-                f"If used, `initial_train_size` must be an integer greater than "
+                f"If `initial_train_size` is an integer, it must be greater than "
                 f"the `window_size` of the forecaster ({forecaster.window_size}) "
-                f"and smaller than the length of `{data_name}` ({data_length})."
+                f"and smaller than the length of `{data_name}` ({data_length}). If "
+                f"it is a date, it must be within this range of the index."
             )
         if initial_train_size + gap >= data_length:
             raise ValueError(
-                f"The combination of initial_train_size {initial_train_size} and "
-                f"gap {gap} cannot be greater than the length of `{data_name}` "
+                f"The total size of `initial_train_size` {initial_train_size} plus "
+                f"`gap` {gap} cannot be greater than the length of `{data_name}` "
                 f"({data_length})."
             )
     else:
@@ -405,7 +421,11 @@ def check_backtesting_input(
         else:
             check_interval(interval=interval, alpha=alpha)
 
-    if not allow_incomplete_fold and data_length - (initial_train_size + gap) < steps:
+    if (
+        not allow_incomplete_fold
+        and initial_train_size is not None
+        and data_length - (initial_train_size + gap) < steps
+    ):        
         raise ValueError(
             f"There is not enough data to evaluate {steps} steps in a single "
             f"fold. Set `allow_incomplete_fold` to `True` to allow incomplete folds.\n"